List categories and tags beneath posts
Instead of just giving a list of a post's tags in the post's footer,
list all the posts for each of the post's categories and tags at the
bottom of the post's page.

Category and tag pages are still supported and will be linked to by
these category and tag listings at the bottom of posts.

The new category_link.html and tag_link.html includes implement
rendering a category or tag's name, linked to the category or tag page
if it exists.

Also tidied up the example tags and categories, applying them to
specific example tagged and categorised posts that're missing from the
main post list, instead of randomy tagging and categorising other
feature posts.