Tweak how category URLs and labels work
Since Jekyll itself puts categories in URLs, it's up to the user to
design category names so that they make pretty URLs. Jekyll won't
downcase category names or replace spaces. So you have to name your
category "example-category" not "Example Category".

If an /example-category/ page exists, though, allow this page's title to
be used as the label when linking to the category, instead of the
category name itself.

With tags it's different since Jekyll doesn't allow those in permalink
URLs. So "Example Tag" with capitals and spaces is fine. The theme
itself converts that to /tags/example-tag/ when finding the tag page.